+++ cmd/zetae2e/bitcoin_address.go
@@ -10,10 +10,8 @@ import (
 
 	"github.com/zeta-chain/zetacore/app"
 	zetae2econfig "github.com/zeta-chain/zetacore/cmd/zetae2e/config"
-	"github.com/zeta-chain/zetacore/cmd/zetae2e/local"
 	"github.com/zeta-chain/zetacore/e2e/config"
 	"github.com/zeta-chain/zetacore/e2e/runner"
-	"github.com/zeta-chain/zetacore/e2e/utils"
 )
 
 const flagPrivKey = "privkey"
@@ -71,8 +69,6 @@ func runBitcoinAddress(cmd *cobra.Command, args []string) error {
 		conf,
 		ethcommon.HexToAddress(evmAddr),
 		conf.Accounts.EVMPrivKey,
-		utils.FungibleAdminName,     // placeholder value, not used
-		local.FungibleAdminMnemonic, // placeholder value, not used
 		logger,
 	)
 	if err != nil {